moved ColorPickerContainer to its own component
